+++ src/main/java/insane96mcp/enhancedai/modules/enderman/getoverhere/GetOverHere.java
@@ -4,21 +4,28 @@ import insane96mcp.enhancedai.EnhancedAI;
 import insane96mcp.enhancedai.modules.Modules;
 import insane96mcp.enhancedai.setup.NBTUtils;
 import insane96mcp.insanelib.base.Feature;
-import insane96mcp.insanelib.base.Label;
 import insane96mcp.insanelib.base.LoadFeature;
 import insane96mcp.insanelib.base.Module;
 import insane96mcp.insanelib.base.config.Config;
+import net.minecraft.core.registries.Registries;
+import net.minecraft.resources.ResourceLocation;
+import net.minecraft.tags.TagKey;
+import net.minecraft.world.entity.EntityType;
 import net.minecraft.world.entity.monster.EnderMan;
 import net.minecraftforge.event.entity.EntityJoinLevelEvent;
 import net.minecraftforge.eventbus.api.SubscribeEvent;
 
-@Label(name = "[Experimental] Get Over Here", description = "Endermen teleport the player near him when can't reach him for a while.")
-@LoadFeature(module = Modules.Ids.ENDERMAN, enabledByDefault = false)
+@LoadFeature(
+        module = Modules.Ids.ENDERMAN,
+        enabledByDefault = false,
+        name = "[Experimental] Get Over Here",
+        description = "Endermen teleport the player near him when can't reach him for a while."
+)
 public class GetOverHere extends Feature {
+    public static final TagKey<EntityType<?>> TELEPORT_BLACKLIST = TagKey.create(Registries.ENTITY_TYPE, ResourceLocation.fromNamespaceAndPath(EnhancedAI.MOD_ID, "get_over_here_teleport_blacklist"));
 
-    public static final String GET_OVER_HERE = EnhancedAI.RESOURCE_PREFIX + "get_over_here";
-    @Config(min = 0d, max = 1d)
-    @Label(name = "Get Over Here Chance", description = "Chance for a enderman to get the Get Over Here AI")
+    public static final ResourceLocation GET_OVER_HERE = EnhancedAI.location("get_over_here");
+    @Config(min = 0d, max = 1d, name = "Get Over Here Chance", description = "Chance for a enderman to get the Get Over Here AI")
     public static Double getOverHereChance = 0.15d;
 
     public GetOverHere(Module module, boolean enabledByDefault, boolean canBeDisabled) {
@@ -32,7 +39,7 @@ public class GetOverHere extends Feature {
                 || !(event.getEntity() instanceof EnderMan enderman))
             return;
 
-        boolean getOverHere = NBTUtils.getBooleanOrPutDefault(enderman.getPersistentData(), GET_OVER_HERE, enderman.getRandom().nextDouble() < getOverHereChance);
+        boolean getOverHere = NBTUtils.getBooleanOrPutDefault(enderman.getPersistentData(), GET_OVER_HERE.toString(), enderman.getRandom().nextDouble() < getOverHereChance);
 
         if (!getOverHere)
             return;